requirements.txt
PyYAML==5.3.1
requests==2.23.0
python-keycloak==0.24.0
rich==9.10.0

